refactor: move to components from fixtures
