+++ packages/web/src/pages/author/[authorAddress].tsx
@@ -18,7 +18,7 @@ import { Avatar as AuthorAvatar } from 'src/components/molecules/Avatar'
 import useWindowDimensions from '../../fixtures/utility/useWindowDimensions'
 import { useGetAccount } from 'src/fixtures/dev-for-apps/hooks'
 import { AvatarProperty } from 'src/components/molecules/AvatarProperty'
-import { Links } from 'src/fixtures/_pages/author/[authorAddress]/Links'
+import { Links } from 'src/components/_pages/author/[authorAddress]/Links'
 import { blueGradient } from 'src/styles/gradient'
 import Link from 'next/link'
 import { useProvider } from 'src/fixtures/wallet/hooks'
