+++ packages/web/src/pages/create/[market]/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useRouter } from 'next/router'
-import { AuthForm } from 'src/fixtures/_pages/create/[market]/AuthForm'
+import { AuthForm } from 'src/components/_pages/create/[market]/AuthForm'
 import { Footer } from 'src/components/organisms/Footer'
 import { Header } from 'src/components/organisms/Header'
 import { Headline } from 'src/components/atoms/Headline'
